Audit round 7: fix FlagDef int overflow and NONE checkbox parity
FlagDef.value changed from int to qint64 — five flag XML files
have values of 2^31 that overflowed 32-bit int. Flag loading now
uses readLong(). NONE checkbox behavior matches Java: always
checked via generic bitwise check, no-op on toggle.

+++ src/core/XmlIO.cpp
@@ -148,7 +148,7 @@ QList<FlagDef> XmlIO::loadFlags(const QString &filePath)
                 xml.readNext();
                 if (!xml.isStartElement()) continue;
                 if (xml.name() == u"name")       f.name = readText(xml);
-                else if (xml.name() == u"value")  f.value = readInt(xml);
+                else if (xml.name() == u"value")  f.value = readLong(xml);
             }
             result.append(f);
         }
